Add positional assessments
