+++ Tests/ChessKitTests/Utilities/SampleGames.swift
@@ -17,7 +17,7 @@ extension Game {
     [Result "1/2-1/2"]
 
     1. e4 $4 e5 $3 2. Nf3 Nc6 3. Bb5 a6 {This opening is called the Ruy Lopez.}
-    4. Ba4 {test comment} Nf6 5. O-O Be7 6. Re1 b5 7. Bb3 d6 8. c3 O-O 9. h3 Nb8 10. d4 Nbd7
+    4. Ba4 {test comment} Nf6 5. O-O Be7 6. Re1 b5 7. Bb3 d6 $135 8. c3 O-O 9. h3 Nb8 10. d4 Nbd7
     11. c4 c6 12. cxb5 axb5 13. Nc3 Bb7 14. Bg5 b4 15. Nb1 h6 16. Bh4 c5 17. dxe5
     Nxe4 18. Bxe7 Qxe7 19. exd6 Qf6 20. Nbd2 Nxd6 21. Nc4 Nxc4 22. Bxc4 Nb6
     23. Ne5 Rae8 24. Bxf7+ Rxf7 25. Nxf7 Rxe1+ 26. Qxe1 Kxf7 27. Qe3 Qg5 28. Qxg5
